connector/requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6       
dnspython==2.3.0      
fastapi==0.95.0       
h11==0.14.0
httptools==0.5.0      
idna==3.4
psycopg2-binary==2.9.6
pydantic==1.10.7      
pymongo==4.3.3
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
watchfiles==0.18.1
websockets==10.4